requirements.txt
Flask==3.1.2
gunicorn==23.0.0
prometheus_client==0.21.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
redis==5.2.1
requests==2.32.5
schedule==1.2.2
python-dotenv==1.2.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0